+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## [1.0.0-rc.3] - 2019-07-29
+### Added
+- Register of shares page
+
+### Fixed
+- Bug with naming css classes in user-movements
+
 ## [1.0.0-rc.2] - 2019-07-26
 ### Added
 - Show transactions history of the each client for bussiness
@@ -63,7 +70,6 @@ for further information about branching and tagging conventions.
   - `OFFLINE_ISSUANCE_WIN_LINK`
   - `OFFLINE_ISSUANCE_MAC_LINK`
   - `OFFLINE_ISSUANCE_SOURCE_LINK`
-
 #### Changed
 - If some of the config keys are empty, the related images and links will not
   be rendered. Affects the following keys:
@@ -76,7 +82,8 @@ for further information about branching and tagging conventions.
 #### Removed
 - `RECOVERY_MODE` config key due to unused anymore
 
-[Unreleased]: https://github.com/tokend/web-client/compare/1.0.0-rc.2...HEAD
+[Unreleased]: https://github.com/tokend/web-client/compare/1.0.0-rc.3...HEAD
+[1.0.0-rc.3]: https://github.com/tokend/web-client/compare/1.0.0-rc.2...1.0.0-rc.3
 [1.0.0-rc.2]: https://github.com/tokend/web-client/compare/1.0.0-rc.1...1.0.0-rc.2
 [1.0.0-rc.1]: https://github.com/tokend/web-client/compare/1.0.0-rc.0...1.0.0-rc.1
 [1.0.0-rc.0]: https://github.com/tokend/web-client/releases/tag/1.0.0-rc.0